# docs/swagger/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
  model.CreateUser:
    properties:
      avatar:
        description: Ảnh đại diện
        type: string
      email:
        description: Địa chỉ email
        type: string
      full_name:
        description: Tên hiển thị
        type: string
      password:
        description: Mật khẩu phải có tối thiểu 8 ký tự
        type: string
      phone:
        description: Số điện thoại
        type: string
      roles:
        description: |-
          Mảng các quyền hạn (Role) của User (ví dụ [3, 5] tức là User có Role Id là 3 và 5).
          Một User có thể có nhiều Role, nếu Null tức là User không có quyền gì
        items:
          type: int32
        type: array
    type: object
  model.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
host: '{{.Host}}'
info:
  contact: {}
  description: '{{.Description}}'
  license: {}
  title: '{{.Title}}'
  version: '{{.Version}}'
paths:
  /test/create-user:
    post:
      description: Tạo User
      parameters:
      - description: Thông tin tạo User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.CreateUser'
          type: object
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
            type: object
      tags:
      - admin
  /test/delete-user/{id}:
    delete:
      description: Xóa User theo ID
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
            type: object
      tags:
      - admin
  /test/get-user/{id}:
    get:
      description: Lấy thông tin User theo ID
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
            type: object
      tags:
      - admin
  /test/get-users:
    get:
      description: Lấy danh sách User
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
            type: object
      tags:
      - admin
  /test/update-user/{id}:
    put:
      description: Cập nhật User theo ID
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HTTPError'
            type: object
      tags:
      - admin
swagger: "2.0"